Added log out button to all pages
