+++ Trial 2.0/index.html
@@ -53,10 +53,10 @@
     <div class="topnav">
       <img src="images\BillLogoUpdated.png" />
       <button class="topnav hidden" id="authorizeBtn">
+        <script type="module" src="authorization_button.js"></script>
         <img id="profile-picture" class="hidden">
           Login
       </button>
-      <!-- <input id="authorizeBtn" type="button" value=" Login "><img id="profile-picture"><br /> -->
     </div>
       
     <section class="main">
