+++ Trial 2.0/sologamescreen.html
@@ -22,7 +22,11 @@
     <a class="active" href='sologamescreen.html'>Solo</a>
 
     <a href="howtoplay.html">How to Play</a>
-    <input id="authorizeBtn" type="button" value=" Login "><br />
+    <button class="topnav hidden" id="authorizeBtn">
+      <script type="module" src="authorization_button.js"></script>
+      <img id="profile-picture" class="hidden">
+        Login
+    </button>
   </div>
 
   <section id="game-section">
